attempted revision of nimble installation to include eigen and svd functions
