+++ packages/nimble/configure
@@ -4219,5 +4219,6 @@ else
   cp inst/CppCode/dists.cpp src
   cp inst/CppCode/RcppUtils.cpp src
   cp inst/CppCode/nimbleGraph.cpp src
+  cp inst/CppCode/EigenTypedefs.cpp src
   (cd inst/CppCode ; ${MAKE} libnimble.a)
 fi
