+++ packages/nimble/src/nimble.cpp
@@ -1,11 +1,11 @@
 #include <Eigen/Dense>
 #include <iostream> // must go before other things because R defines a "length" macro
+#include <nimble/EigenTypedefs.h>
 #include <nimble/RcppUtils.h>
-#include "nimble/RcppNimbleUtils.h"
+//#include "nimble/RcppNimbleUtils.h"
 //#include <nimble/ModelClassUtils.h>
 //#include <nimble/accessorClasses.h>
 #include <nimble/nimbleGraph.h>
-//#include <nimble/EigenTypedefs.h>
 #include <nimble/dists.h>
 
 #include <R_ext/Rdynload.h>
